requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.4.22
Django==6.0.4
djangorestframework==3.17.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
Markdown==3.10.2
ollama==0.6.1
pydantic==2.13.3
pydantic_core==2.46.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
